refactor(adk): improve turn loop structure
